New simulation completion test
Added a new condition in _load_plumed() function to check the length of the simulation data being loaded. This will alert the user in case a simulation has been terminated because of queue time limits which will be marked as a sucessful process and creates a .done file although the simulation is not actually compelete.