New pre-optimization step added.
Added a new step before running acpype that uses openbabel and gaff ff to pre-optimize the structure generated with pymol. This is necessary for sequence that have large side chains because pymol does not check for clashes after chiral transformation to D-AAs and in most cases the resulting structure is not physical and acpype fails to determine the molecular graph. A short optimization step should take care of particles that are too close to each other.