New procedure for failed runs and warning for FE calculations.
fixed backup procedure for HILLS files after resuming a previously failed run. Also added a RuntimeError that is raised if the run exits with code 0 but it is not actually finished (checks the existance of md.gro file).
Block analysis is now inside of a try/except block for cases where the system is stuck in the bound state (or doesn't bind) for the duration of a full block. These blocks are DISCARDED and a warning is printed to notify the user.